+++ 07/curtis_part2.cpp
@@ -11,8 +11,8 @@ using std::vector;
 
 enum ops { ADD, MUL, CAT };
 
-long long evaluate(const vector<long long> &nums, const vector<ops> &operators) {
-    long long acc = nums.at(0);
+long evaluate(const vector<long> &nums, const vector<ops> &operators) {
+    long acc = nums.at(0);
     auto current_num = nums.begin();
     auto current_op = operators.begin();
     while (++current_num < nums.end()) {
@@ -74,7 +74,7 @@ struct OpGenerator {
     }
 };
 
-bool works(const vector<long long> &nums, long long value) {
+bool works(const vector<long> &nums, long value) {
     OpGenerator gen(nums.size() - 1);
     vector<ops> operators;
     
@@ -90,9 +90,9 @@ int main (int argc, char *argv[]) {
     std::ifstream input(argv[1]);
     std::string s;
 
-    vector<vector<long long>> numbers;
+    vector<vector<long>> numbers;
     numbers.reserve(850);
-    vector<long long> values;
+    vector<long> values;
     values.reserve(850);
 
     while (std::getline(input, s)) {
@@ -108,14 +108,14 @@ int main (int argc, char *argv[]) {
         numbers.back().pop_back();
     }
 
-    long long acc = 0;
+    long acc = 0;
     for (size_t x = 0; x < numbers.size(); ++x) {
         if (works(numbers.at(x), values.at(x))) {
             acc += values.at(x);
         }
     }
 
-    std::printf("Solution: %lld!\n", acc);
+    std::printf("Solution: %ld!\n", acc);
 
     return 0;
 }
